-Added search functionality (still need to figure out how to link to it) -Added the src directory back in its slimmed-down form -Setup checks for Rake now, since apparently no errors were thrown in its absence

+++ setup.rb
@@ -30,6 +30,8 @@ else
 	quiet = "--quiet"
 end
 
+abort "Rake required for compilation - please use 'gem install rake' to install. You make uninstall after successfully building Sqwee." unless system("rake --version")
+
 FileUtils.rm_rf("bin/*", {:secure => true})
 FileUtils.rm_rf("extlib/*", {:secure => true})
